+++ green-plate/src/main/java/com/team404x/greenplate/orders/model/response/OrderUserSearchDetailRes.java
@@ -11,7 +11,7 @@ public class OrderUserSearchDetailRes {
     private Long order_id;
     private Long orderDetail_id;
     private LocalDateTime order_date;
-    private Long price;
+    private Integer price;
     private Integer cnt;
     private String order_state;
     private String zipCode;
